Fixed DXSDK errors + added resources
The compiler errors I had been encountering up until now were due to
conflicts between the legacy DirectX SDK and the modern Windows SDK that
I had installed. In order to prevent this, I've tweaked the generators
so that the modern equivalents will be used instead.

Also added some shaders and an icon for the project.