Removed erroneous printf statements
When running the application through the CLI, a number of leftover input
debug print statements were ruining the output.

Also played around a little more with lights - Reached the conclusion
that, for more advanced lighting techniques, an overhaul to the renderer
will be necessary. (As a stretch goal, of course)