Fixed Blinn Shading
Implemented and fixed the blinn shader's output.
There were a number of issues, mainly stemming from a bug in the CPU side loading code, wherein the normal map was incorrectly being assigned to the specular map, and within the shader, the specular map was being assigned to the normal map.